requirements.txt
Scrapy>=2.11.0      # web scraping
Django
tqdm                # progress bar

selenium
requests
2captcha-python
beautifulsoup4

psycopg2-binary         # postgres connection
python-dotenv           # get env credentials
gunicorn

django-debug-toolbar
boto3
sentry-sdk[django]

redis                   # caching
celery[redis]           # caching
django-celery-beat

crochet
openai
djangorestframework
django-filter
djangorestframework-api-key==3.*

pytest
pytest-django